pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "japan-pr-navigator"
version = "2.0.0"
description = "MCP server for Japan Permanent Residence application guidance"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6.0",
    "mypy>=1.11",
    "types-PyYAML>=6.0",
    "types-beautifulsoup4>=4.12",
]

[tool.hatch.build.targets.wheel]
packages = ["src/japan_pr_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["src/japan_pr_mcp"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
